Move DeleteDeviceButton to Buttons folder & update imports
